pyproject.toml
[project]
name = "rag-council"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "datasets>=4.8.4",
    "gradio>=6.9.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "pypandoc>=1.17",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "ragas>=0.4.3",
    "unstructured>=0.21.5",
]
